+++ source/FlexTable/FlexColumn.js
@@ -1,76 +1,8 @@
 /** @flow */
-import React, { Component, PropTypes } from 'react'
-import type { CellDataGetterParams, CellRendererParams } from './types'
-import SortIndicator from './SortIndicator'
-
-/**
- * Default cell renderer that displays an attribute as a simple string
- * You should override the column's cellRenderer if your data is some other type of object.
- */
-export function defaultCellRenderer ({
-  cellData,
-  cellDataKey,
-  columnData,
-  rowData,
-  rowIndex
-}: CellRendererParams): string {
-  if (cellData === null || cellData === undefined) {
-    return ''
-  } else {
-    return String(cellData)
-  }
-}
-
-/**
- * Default accessor for returning a cell value for a given attribute.
- * This function expects to operate on either a vanilla Object or an Immutable Map.
- * You should override the column's cellDataGetter if your data is some other type of object.
- */
-export function defaultCellDataGetter ({
-  columnData,
-  dataKey,
-  rowData
-}: CellDataGetterParams) {
-  if (rowData.get instanceof Function) {
-    return rowData.get(dataKey)
-  } else {
-    return rowData[dataKey]
-  }
-}
-
-/**
- * Default table header renderer.
- */
-export function defaultHeaderRenderer ({
-  columnData,
-  dataKey,
-  disableSort,
-  label,
-  sortBy,
-  sortDirection
-}) {
-  const showSortIndicator = sortBy === dataKey
-  const children = [
-    <div
-      className='FlexTable__headerTruncatedText'
-      key='label'
-      title={label}
-    >
-      {label}
-    </div>
-  ]
-
-  if (showSortIndicator) {
-    children.push(
-      <SortIndicator
-        key='SortIndicator'
-        sortDirection={sortDirection}
-      />
-    )
-  }
-
-  return children
-}
+import { Component, PropTypes } from 'react'
+import defaultHeaderRenderer from './defaultHeaderRenderer'
+import defaultCellRenderer from './defaultCellRenderer'
+import defaultCellDataGetter from './defaultCellDataGetter'
 
 /**
  * Describes the header and cell contents of a table column.
